restaurant-backend/src/main/java/com/restaurant/service/impl/ProductServiceImpl.java
package com.restaurant.service.impl;

import com.restaurant.entity.Category;
import com.restaurant.entity.Product;
import com.restaurant.repository.ProductRepository;
import com.restaurant.service.ProductService;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductServiceImpl implements ProductService {


    private final ProductRepository productRepository;

    public ProductServiceImpl(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }


    @Override
    public Product createProduct(Product product) {
        return productRepository.save(product);
    }

    @Override
    public Product getProductById(Long id) {
        return productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
    }

    @Override
    public List<Product> getAllProducts() {
        return productRepository.findAll();
    }

    @Override
    public List<Product> getAllProductsByCategory(Category category) {
        return productRepository.findAllByCategory(category);
    }
}
